session_4/requirements.txt
attrs==22.2.0
beaker-pyteal==0.5.2
cffi==1.15.1
docstring-parser==0.14.1
exceptiongroup==1.1.0
iniconfig==1.1.1
msgpack==1.0.4
packaging==22.0
pluggy==1.0.0
py-algorand-sdk==2.0.0
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
pyteal==0.22.0
pytest==7.2.0
semantic-version==2.10.0
tomli==2.0.1